backend/alembic/env.py
import asyncio
import os
from logging.config import fileConfig

from sqlalchemy import pool
from sqlalchemy.engine import Connection
from sqlalchemy.ext.asyncio import async_engine_from_config

from alembic import context
from dotenv import load_dotenv

load_dotenv(override=True)

# Alembic Config object
config = context.config

# Set the database URL from env (overrides alembic.ini placeholder)
config.set_main_option("sqlalchemy.url", os.getenv("DATABASE_URL", ""))

# Logging
if config.config_file_name is not None:
    fileConfig(config.config_file_name)

# Import ALL models so Alembic can detect them
from backend.database import Base  # noqa: E402
from backend.models.user import User  # noqa: E402, F401
from backend.models.guest_session import GuestSession  # noqa: E402, F401
from backend.models.resume import Resume  # noqa: E402, F401
from backend.models.job import JobDescription  # noqa: E402, F401
from backend.models.analysis import Analysis  # noqa: E402, F401
from backend.models.password_reset import PasswordResetToken  # noqa: E402, F401

target_metadata = Base.metadata


def run_migrations_offline() -> None:
    url = config.get_main_option("sqlalchemy.url")
    context.configure(
        url=url,
        target_metadata=target_metadata,
        literal_binds=True,
        dialect_opts={"paramstyle": "named"},
    )
    with context.begin_transaction():
        context.run_migrations()


def do_run_migrations(connection: Connection) -> None:
    context.configure(connection=connection, target_metadata=target_metadata)
    with context.begin_transaction():
        context.run_migrations()


async def run_async_migrations() -> None:
    connectable = async_engine_from_config(
        config.get_section(config.config_ini_section, {}),
        prefix="sqlalchemy.",
        poolclass=pool.NullPool,
    )
    async with connectable.connect() as connection:
        await connection.run_sync(do_run_migrations)
    await connectable.dispose()


def run_migrations_online() -> None:
    asyncio.run(run_async_migrations())


if context.is_offline_mode():
    run_migrations_offline()
else:
    run_migrations_online()